app/requirements.txt
pandas==1.3.5
scikit-learn==1.0.2
joblib==1.1.0
paho-mqtt==1.6.1